Move formats.go (rename awk.go) to cli package, rename and document function
